requirements.txt
bootstrapper==1.0.0
diff-match-patch==20121119
django-import-export==0.5.1
django-summernote==0.8.7
et-xmlfile==1.0.1
jdcal==1.3
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
Pillow==4.1.1
python-http-client==2.2.1
pytz==2017.2
PyYAML==3.12
sendgrid==4.2.1
tablib==0.12.1
unicodecsv==0.14.1
virtualenv==15.1.0
xlrd==1.1.0
xlwt==1.3.0
